+++ CHANGELOG.md
@@ -5,5 +5,6 @@
 - Better stack traces (available through `DebugTrace.stackTrace()`)
 - Threads now have a changeable name
 - Turned a few array parameters into varargs where possibly (notably `LuaValue.varargsof()`)
+- Added a user-settable uncaught exception handler for Lua threads (`LuaRunState.setExceptionHandler`)
 - fix: The 'call' debug hook was wrongly called with `__call` instead of `call`.
 - performance: stack/upValue arrays are sized smaller and reused to reduce garbage collection pressure.